make use of various data types consistent and use posix types if possible
this cleans up the mess with long vs. unsigned long either directly
or via various typedefs for all kinds of variables that store time data.
we now use time_t, as far as the compiler could detect (the rest will be "long").
similarly, use uid_t and gid_t consistently.

+++ src/moab/MLimit.c
@@ -66,7 +66,7 @@ int MLimitEnforceAll(
     if ((JobWCX >= 0) &&
         (J->WCLimit > 0) &&
         (MSched.Time > J->StartTime) &&
-       ((unsigned long)(MSched.Time - J->StartTime) > (J->WCLimit + J->SWallTime + JobWCX)))
+       ((MSched.Time - J->StartTime) > (J->WCLimit + J->SWallTime + JobWCX)))
       {
       DBG(2,fCORE) DPrint("ALERT:    job '%s' in state '%s' has exceeded its wallclock limit (%ld+S:%ld) by %s (job will be cancelled)\n",
         J->Name,
